Add files via upload
We have been asked to find out insights from the 'Amazon Fine Foods Review' data. Our aim is to do text classification using Deep Learning on the reviews. This dataset has various features but for this project, we will only need the reviews(text) as training data. This is a supervised Learning task so you will use Score as the label while training. You have to make a classification model that will rate the review at a scale from 1 to 5.